rename virtual_keyboard to virtualKeyboard
